CCCP-227, refactor: scaffold
