+++ rbclib/periodic/heartbeat.py
@@ -5,10 +5,10 @@ from chainpy.eventbridge.periodiceventabc import PeriodicEventABC
 from chainpy.eventbridge.utils import timestamp_msec
 from chainpy.logger import global_logger
 
-from rbclib.utils import is_pulsed_hear_beat
-from rbclib.chainevents import NoneParams
+from rbclib.chain_events import NoneParams
 from rbclib.metric import PrometheusExporterRelayer
-from rbclib.switchable_enum import chain_primitives
+from rbclib.primitives.relay_chain import chain_primitives
+from rbclib.utils import is_pulsed_hear_beat
 from relayer.global_config import relayer_config_global
 from relayer.relayer import Relayer
